refactor: organize platforms directive models and arbitraries
