+++ platforms/guardoni/tsconfig.json
@@ -27,7 +27,9 @@
       "@project/*": ["./guardoni/tx-automate/project/*"],
       "@platform/*": ["./guardoni/tx-automate/platform/*"],
       "@storage/*": ["./guardoni/tx-automate/storage/*"],
-      "@util/*": ["./guardoni/tx-automate/util/*"]
+      "@util/*": ["./guardoni/tx-automate/util/*"],
+      "@yttrex/shared/*": ["../../yttrex/shared/src/*"],
+      "@tktrex/shared/*": ["../../tktrex/shared/src/*"]
     }
   },
   "include": ["./src", "__mocks__", "__tests__"],
@@ -46,6 +48,12 @@
   "references": [
     {
       "path": "../../packages/shared"
+    },
+    {
+      "path": "../yttrex/shared"
+    },
+    {
+      "path": "../tktrex/shared"
     }
   ]
 }
